UR10e_anamoly/data-analytics/requirements.txt
pip
paho-mqtt
numpy
scikit-learn
torch